CNV-71648: hide system namespaces in ssh modal
