+++ src/utils/resources/namespace/constants.ts
@@ -1,0 +1,2 @@
+export const SYSTEM_NAMESPACES_PREFIX = ['kube-', 'openshift-', 'kubernetes-'];
+export const SYSTEM_NAMESPACES = ['default', 'openshift'];
