+++ src/utils/resources/namespace/helper.ts
@@ -1,0 +1,8 @@
+import { SYSTEM_NAMESPACES, SYSTEM_NAMESPACES_PREFIX } from './constants';
+
+export const isSystemNamespace = (projectName: string) => {
+  const startsWithNamespace = SYSTEM_NAMESPACES_PREFIX.some((ns) => projectName.startsWith(ns));
+  const isNamespace = SYSTEM_NAMESPACES.includes(projectName);
+
+  return startsWithNamespace || isNamespace;
+};
