+++ src/views/virtualmachines/tree/utils/constants.ts
@@ -9,8 +9,6 @@ export const OPEN_DRAWER_SIZE = '400px';
 export const CLOSED_DRAWER_SIZE = '30px';
 export const PANEL_WIDTH_PROPERTY = '--pf-v6-c-drawer__panel--md--FlexBasis';
 
-export const SYSTEM_NAMESPACES_PREFIX = ['kube-', 'openshift-', 'kubernetes-'];
-export const SYSTEM_NAMESPACES = ['default', 'openshift'];
 export const SHOW_EMPTY_PROJECTS_KEY = 'showEmptyProjects';
 export const TREE_VIEW_LAST_WIDTH = 'treeViewLastWidth';
 export const SHOW_TREE_VIEW = 'showTreeView';
